requirements.txt
annotated-types==0.7.0
anyio==4.12.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.1.8
distro==1.9.0
exceptiongroup==1.3.1
groq==0.37.1
gTTS==2.5.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jiter==0.12.0
numpy==2.2.6
openai==2.9.0
PyAudio==0.2.14
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
pygame==2.6.1
python-dotenv==1.2.1
requests==2.32.5
sniffio==1.3.1
sounddevice==0.5.3
SpeechRecognition==3.14.4
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.1
